tlog does not own sssd.conf - so use ini_file to manage it
tlog does not own sssd.conf, so we cannot use a template for it,
since that may wipe out changes someone else made.  Instead, use
`ini_file` (and add a requirement on `community.general`) to manage
the necessary entries in the file.

In addition, this only needs to be done on rhel9+ and fedora, so on
rhel8, skip this step.