fix: Skip runtime operations in non-systemd environments
Cause: the role and tests were previously assuming that the system was
booted with systemd and could run/restart services. But this is not the
case when running the role during container builds.

Consequence: The role did not work during bootc container builds.

Fix: Detect if the system is booted (with systemd), and skip service
restarting and tests_default_wrapper.yml if not.

Result: The role now works during container builds.

https://issues.redhat.com/browse/RHEL-89561