fix: Set __is_system_running in check mode
Cause: `systemctl is-system-running` was not called in `--check` mode.

Consequence: The "Determine if system is booted with systemd" step
failed in check mode as the `__is_system_running` variable was not
populated.

Fix: Force calling `systemctl is-system-running` in check mode. It does
not modify the system and the outcome is very influential for what the
role does.