fix: Fix assertions to have a boolean type
Cause: Ansible 2.19 requires explicit boolean types for assertions.

Consequence: The test failed with Ansible 2.19 with "Conditional result
was '62' of type 'int', which evaluates to True. Conditionals must have
a boolean result."

Fix: find() does not actually have a result which is compatible with
boolean coercion, as 0 means "string found at the start", and "not
found" results in -1 (which is `True`). We could compare the find()
result explicitly, but let's instead use the `in` operator to make this
more readable.

Also rewrite these assertions to use Python string concatenation. This
fixes

> [WARNING]: conditional statements should not include jinja2 templating delimiters such as {{ }} or {% %}. Found: 'scope={{
tlog_scope_sssd }}' in __tlog_conf_content.stdout

Likewise, `regex_search()` returns a string (possibly empty for
non-matches) or `none`, so explicitly check the length to convert to a
bool.